requirements.txt
flask
flask-cors
transformers
torch
groq
python-dotenv
